// web-app/src/generated/model/orderItemDto.ts
/**
 * Api
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { ProductDto } from './productDto';


export interface OrderItemDto { 
    orderId?: number;
    quantity?: number;
    price?: number;
    product?: ProductDto;
}