// web-app/src/generated/model/orderSummaryDto.ts
/**
 * Api
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { OrderItemDto } from './orderItemDto';


export interface OrderSummaryDto { 
    orderId?: number;
    companyName?: string | null;
    description?: string | null;
    orderTotal?: number;
    orderProducts?: Array<OrderItemDto> | null;
}